requirements.txt
python-dotenv
pandas>=2.0.0
pyarrow>=14.0.0
polars>=0.20.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0
kaleido>=0.2.0
